Change whitespace handling
By changing the white-space rule to pre, a single space will not be
spontaneously removed. This means we can calculate width etc of spaces
without worrying about this clipping.

Since we do not worry about the clipping any more, spaces will be shown
as intended and will also copy/paste as expected.

[Ticket: 1]