Use functional components
This commit is WAY too big for a single commit. not good. but here are
the changes.

Changed to use react functional components with hooks instead of class
components with state etc. Makes the code much easier to read, and
understand. Also, changed the edit implementation slightly to make it
more readable as well.

One big change now, is that the character components, will remember the
previous character on their own when it updates and change accordingly.
This means that the edit function no longer has to return a move with
the previous character, as that is already remembered.

This has caused a single problem though. When the character is updated,
there is a tiny moment where the component is rendered with the new
character, but the effect hook that manages whether the character has
changed is called after the render, causing a tiny flicker where the
characters have changed but the class governing the animation has not.

Also, added rules in webpack to make sure that we don't pack react in
with the component.

My main problem is that moving from class to functional components was
a big jump, and I didn't want to commit code that didn't work. I now
know that this was a mistake. This commit is HUGE and there are no
intermediate commits that I can see how the code got from a to b.

If I learned my lesson I guess this will be the biggest commit I write
in a while. Here's hoping for better commits in future!