Change character display
Previously when the character was updated, this would occur

1. The effect hooks would be triggered since the prop was different.
2. The component would be displayed with the new character.
3. The effect hooks would run and update the state to trigger the
  animation.
4. The component would render again with the new character fading in and
  the old character fading out.

This caused a problem, because for the short delay between 2. and 4.,
the component would show the next character without a transition,
causing the display to flicker.

By storing the display characters separately in state, we can still
display the previous characters until the animation will trigger,
removing the flicker.