Fix inequality
The widen delay is supposed to make sure that if the character will
grow, then the width changes before the fade, and if the character will
shrink, then the width changes after the fade.

The inequality was around the wrong way, so the character would shrink,
fade and then grow.